+++ R/lsm_l_gyrate_cv.R
@@ -5,6 +5,8 @@
 #' @param landscape Raster* Layer, Stack, Brick or a list of rasterLayers.
 #' @param directions The number of directions in which patches should be
 #' connected: 4 (rook's case) or 8 (queen's case).
+#' @param cell_center If true, the coordinates of the centroid are forced to be
+#' a cell center within the patch.
 #'
 #' @details
 #' \deqn{GYRATE_{CV} = cv(GYRATE[patch_{ij}])}
@@ -50,15 +52,17 @@
 #' in fragmented landscapes. Conservation ecology, 1(1).
 #'
 #' @export
-lsm_l_gyrate_cv <- function(landscape, directions) UseMethod("lsm_l_gyrate_cv")
+lsm_l_gyrate_cv <- function(landscape, directions, cell_center) UseMethod("lsm_l_gyrate_cv")
 
 #' @name lsm_l_gyrate_cv
 #' @export
-lsm_l_gyrate_cv.RasterLayer <- function(landscape, directions = 8) {
+lsm_l_gyrate_cv.RasterLayer <- function(landscape,
+                                        directions = 8, cell_center = FALSE) {
 
     result <- lapply(X = raster::as.list(landscape),
                      FUN = lsm_l_gyrate_cv_calc,
-                     directions = directions)
+                     directions = directions,
+                     cell_center = cell_center)
 
     layer <- rep(seq_along(result),
                  vapply(result, nrow, FUN.VALUE = integer(1)))
@@ -70,11 +74,13 @@ lsm_l_gyrate_cv.RasterLayer <- function(landscape, directions = 8) {
 
 #' @name lsm_l_gyrate_cv
 #' @export
-lsm_l_gyrate_cv.RasterStack <- function(landscape, directions = 8) {
+lsm_l_gyrate_cv.RasterStack <- function(landscape,
+                                        directions = 8, cell_center = FALSE) {
 
     result <- lapply(X = raster::as.list(landscape),
                      FUN = lsm_l_gyrate_cv_calc,
-                     directions = directions)
+                     directions = directions,
+                     cell_center = cell_center)
 
     layer <- rep(seq_along(result),
                  vapply(result, nrow, FUN.VALUE = integer(1)))
@@ -86,11 +92,13 @@ lsm_l_gyrate_cv.RasterStack <- function(landscape, directions = 8) {
 
 #' @name lsm_l_gyrate_cv
 #' @export
-lsm_l_gyrate_cv.RasterBrick <- function(landscape, directions = 8) {
+lsm_l_gyrate_cv.RasterBrick <- function(landscape,
+                                        directions = 8, cell_center = FALSE) {
 
     result <- lapply(X = raster::as.list(landscape),
                      FUN = lsm_l_gyrate_cv_calc,
-                     directions = directions)
+                     directions = directions,
+                     cell_center = cell_center)
 
     layer <- rep(seq_along(result),
                  vapply(result, nrow, FUN.VALUE = integer(1)))
@@ -102,13 +110,15 @@ lsm_l_gyrate_cv.RasterBrick <- function(landscape, directions = 8) {
 
 #' @name lsm_l_gyrate_cv
 #' @export
-lsm_l_gyrate_cv.stars <- function(landscape, directions = 8) {
+lsm_l_gyrate_cv.stars <- function(landscape,
+                                  directions = 8, cell_center = FALSE) {
 
     landscape <- methods::as(landscape, "Raster")
 
     result <- lapply(X = raster::as.list(landscape),
                      FUN = lsm_l_gyrate_cv_calc,
-                     directions = directions)
+                     directions = directions,
+                     cell_center = cell_center)
 
     layer <- rep(seq_along(result),
                  vapply(result, nrow, FUN.VALUE = integer(1)))
@@ -120,11 +130,13 @@ lsm_l_gyrate_cv.stars <- function(landscape, directions = 8) {
 
 #' @name lsm_l_gyrate_cv
 #' @export
-lsm_l_gyrate_cv.list <- function(landscape, directions = 8) {
+lsm_l_gyrate_cv.list <- function(landscape,
+                                 directions = 8, cell_center = FALSE) {
 
     result <- lapply(X = landscape,
                      FUN = lsm_l_gyrate_cv_calc,
-                     directions = directions)
+                     directions = directions,
+                     cell_center = cell_center)
 
     layer <- rep(seq_along(result),
                  vapply(result, nrow, FUN.VALUE = integer(1)))
@@ -134,11 +146,12 @@ lsm_l_gyrate_cv.list <- function(landscape, directions = 8) {
     tibble::add_column(result, layer, .before = TRUE)
 }
 
-lsm_l_gyrate_cv_calc <- function(landscape, directions,
+lsm_l_gyrate_cv_calc <- function(landscape, directions, cell_center,
                                  points = NULL) {
 
     gyrate_patch <- lsm_p_gyrate_calc(landscape,
                                       directions = directions,
+                                      cell_center = cell_center,
                                       points = points)
 
     # all values NA